Move logic sellStock, data invoice to model transaction, and fix bug view sell stock, view sell history
