+++ controllers/TransactionController.js
@@ -25,10 +25,6 @@ class TransactionController {
             const type = "Buy"
 
             const user = await User.reduceBalanceUser(UserId, Number(totalPrice))
-
-            req.session.UserId = user.id
-            req.session.role = user.role
-            req.session.username = user.username
             req.session.balance = user.balance
             
             await Transaction.create({
@@ -55,7 +51,7 @@ class TransactionController {
             const { StockId } = req.params
             const { username, balance, UserId} = req.session
 
-            const stock = await Stock.findOne({
+            let stock = await Stock.findOne({
                 include:{
                     model: Transaction,
                     where: {
@@ -69,7 +65,8 @@ class TransactionController {
                 order: [[{ model: Transaction }, 'createdAt', 'DESC']]
             })
             if(!stock){
-                res.redirect("/portfolio")
+                stock = await Stock.findByPk(StockId)
+                stock.Transactions = []
             }
 
             res.render("transaction-sell" , {username, balance, formatRupiah, stock, formatDateLocal})
@@ -84,33 +81,11 @@ class TransactionController {
             const { TransactionId } = req.params
             const { username, balance, UserId} = req.session
             
-            const transaction = await Transaction.findByPk(TransactionId,{
-                include: {
-                    model: Stock
-                }
-            })
-
-            let sellPrice = transaction.Stock.currentPrice
-            let totalSellPrice = sellPrice * transaction.qty * 100
-            const profit = totalSellPrice - transaction.totalPrice
-            let profitPersen = ( totalSellPrice - transaction.totalPrice) / totalSellPrice * 100
-            profitPersen = profitPersen.toFixed(2)
-
-            // res.send(totalSellPrice)
-            await transaction.update({
-                sellPrice,
-                profit,
-                profitPersen,
-                type: "Sell",
-                totalSellPrice
-            })
-
+            const transaction = await Transaction.sellStock(TransactionId)
+            
             await Portfolio.createOrUpdatePortfolio(transaction.Stock.id, UserId)
 
-            const user = await User.addBalanceUser(UserId, Number(totalSellPrice))
-            req.session.UserId = user.id
-            req.session.role = user.role
-            req.session.username = user.username
+            const user = await User.addBalanceUser(UserId, Number(transaction.totalSellPrice))
             req.session.balance = user.balance
 
             res.redirect(`/transaction/sell/history/${transaction.Stock.id}`)
@@ -124,7 +99,7 @@ class TransactionController {
             const { StockId } = req.params
             const { username, balance, UserId} = req.session
 
-            const stock = await Stock.findOne({
+            let stock = await Stock.findOne({
                 include:{
                     model: Transaction,
                     where: {
@@ -137,6 +112,10 @@ class TransactionController {
                 },
                 order: [[{ model: Transaction }, 'createdAt', 'DESC']]
             })
+            if(!stock){
+                stock = await Stock.findByPk(StockId)
+                stock.Transactions = []
+            }
             res.render("transaction-history-sell" , {username, balance, formatRupiah, stock, formatDateLocal})
 
         } catch (error) {
@@ -161,7 +140,7 @@ class TransactionController {
                 }
             })
 
-            const data = await TransactionController.invoiceData(user, transaction)
+            const data = await Transaction.invoiceData(user, transaction)
             
             const result = await easyinvoice.createInvoice(data);
             const fileName = `Invoice ${transaction.Stock.code} Stock, ${formatDateLocal(transaction.createdAt)}.pdf`
@@ -172,83 +151,6 @@ class TransactionController {
             
         } catch (error) {
             res.send(error)
-        }
-    }
-
-    static async invoiceData(user, transaction){
-        try {
-            
-            let client = user.username
-            if(user.Profile){
-                client = user.Profile.name
-            }
-
-            // seharusnya bisa di getter
-            const dateBuy = formatDateLocal(transaction.createdAt)
-            const date = new Date(transaction.createdAt).toLocaleString('id-ID',{year:'numeric', month:'2-digit', day:'2-digit'}).split("/"); // dd-mm-yyyy
-            const [day,month,year] = date
-            const code = `${year}-${transaction.id}-${transaction.Stock.code}`
-
-            return {
-                // karena gk pakai env api key akan di hapus tgl 9-jun-2025
-                apiKey: "Whlk6m70k6DyTXKdwPpxIoLcMQDhaCwf0kvcitgLmiehafUN7HpX0tWyxMbiIx2a",
-                mode: "development",
-                // "customize": {
-                //     "template": "SGVsbG8gd29ybGQh" // Must be base64 encoded html. This example contains 'Hello World!' in base64
-                // },
-                sender: {
-                    company: 'Modal Receh',
-                    address: 'Keputih',
-                    zip: '60111',
-                    city: 'Surabaya',
-                    country: 'Indonesia'
-                },
-                client: {
-                    company: "Client Name: "+client,
-                },
-                information: {
-                    number: code,
-                    date: dateBuy,
-                },
-                products: [
-                    {
-                        quantity: (transaction.qty*100).toString(),
-                        description: `${transaction.Stock.code} Stocks`,
-                        tax: 0,
-                        price: transaction.price
-                    },
-                ],
-                'bottom-notice': '',
-                settings: {
-                    currency: 'IDR', // See documentation 'Locales and Currency' for more info. Leave empty for no currency.
-                    "locale": "id-ID", // Defaults to en-US, used for number formatting (see docs)
-                    // "margin-top": 25, // Default to 25
-                    // "margin-right": 25, // Default to 25
-                    // "margin-left": 25, // Default to 25
-                    // "margin-bottom": 25, // Default to 25
-                    // "format": "Letter", // Defaults to A4,
-                    // "height": "1000px", // allowed units: mm, cm, in, px
-                        // "width": "500px", // allowed units: mm, cm, in, px
-                        // "orientation": "landscape", // portrait or landscape, defaults to portrait
-                },
-                // Used for translating the headers to your preferred language
-                // Defaults to English. Below example is translated to Dutch
-                "translate": {
-                //     "invoice": "FACTUUR",
-                //     "number": "Nummer",
-                //     "date": "Datum",
-                //     "due-date": "Verloopdatum",
-                //     "subtotal": "Subtotaal",
-                    "products": "Saham",
-                    "quantity": "Jumlah Saham (Lembar)",
-                    "price": "Harga",
-                //     "product-total": "Totaal",
-                //     "total": "Totaal"
-                //		 "vat": "btw"
-                },
-            };
-        } catch (error) {
-            throw error
         }
     }
